+++ sdk/keyvault/keyvault-certificates/test/public/mergeAndImport.spec.ts
@@ -11,10 +11,10 @@ import { ClientSecretCredential } from "@azure/identity";
 
 import { CertificateClient } from "../../src";
 import { base64ToUint8Array, stringToUint8Array } from "../../src/utils";
-import { testPollerProperties } from "../utils/recorderUtils";
-import { authenticate } from "../utils/testAuthentication";
-import { getServiceVersion } from "../utils/utils.common";
-import TestClient from "../utils/testClient";
+import { testPollerProperties } from "./utils/recorderUtils";
+import { authenticate } from "./utils/testAuthentication";
+import { getServiceVersion } from "./utils/common";
+import TestClient from "./utils/testClient";
 
 describe("Certificates client - merge and import certificates", () => {
   const prefix = `merge${env.CERTIFICATE_NAME || "CertificateName"}`;