+++ sdk/test-utils/recorder/src/recorder.ts
@@ -3,9 +3,7 @@
 
 import {
   createDefaultHttpClient,
-  createPipelineRequest,
   HttpClient,
-  HttpMethods,
   Pipeline,
   PipelinePolicy,
   PipelineRequest,
@@ -27,7 +25,7 @@ import { Test } from "mocha";
 import { sessionFilePath } from "./utils/sessionFilePath";
 import { SanitizerOptions } from "./utils/utils";
 import { paths } from "./utils/paths";
-import { Sanitizer } from "./sanitizer";
+import { addSanitizers, transformsInfo } from "./sanitizer";
 import { handleEnvSetup } from "./utils/envSetupForPlayback";
 import { Matcher, setMatcher } from "./matcher";
 import {
@@ -37,6 +35,7 @@ import {
   WebResource,
   WebResourceLike,
 } from "@azure/core-http";
+import { createRecordingRequest } from "./utils/createRecordingRequest";
 
 /**
  * This client manages the recorder life cycle and interacts with the proxy-tool to do the recording,
@@ -55,7 +54,6 @@ export class Recorder {
   private stateManager = new RecordingStateManager();
   private httpClient?: HttpClient;
   private sessionFile?: string;
-  private sanitizer?: Sanitizer;
   private variables: Record<string, string>;
 
   constructor(private testContext?: Test | undefined) {
@@ -68,7 +66,6 @@ export class Recorder {
           "Unable to determine the recording file path, testContext provided is not defined."
         );
       }
-      this.sanitizer = new Sanitizer(this.url, this.httpClient);
     }
     this.variables = {};
   }
@@ -113,8 +110,12 @@ export class Recorder {
    */
   async addSanitizers(options: SanitizerOptions): Promise<void> {
     // If check needed because we only sanitize when the recording is being generated, and we need a recording to apply the sanitizers on.
-    if (isRecordMode() && ensureExistence(this.sanitizer, "this.sanitizer")) {
-      return this.sanitizer.addSanitizers(options);
+    if (
+      isRecordMode() &&
+      ensureExistence(this.httpClient, "this.httpClient") &&
+      ensureExistence(this.recordingId, "this.recordingId")
+    ) {
+      return addSanitizers(this.httpClient, this.url, this.recordingId, options);
     }
   }
 
@@ -135,7 +136,7 @@ export class Recorder {
       const startUri = `${this.url}${isPlaybackMode() ? paths.playback : paths.record}${
         paths.start
       }`;
-      const req = this._createRecordingRequest(startUri);
+      const req = createRecordingRequest(startUri, this.sessionFile, this.recordingId);
 
       if (ensureExistence(this.httpClient, "TestProxyHttpClient.httpClient")) {
         const rsp = await this.httpClient.sendRequest({
@@ -153,12 +154,14 @@ export class Recorder {
         if (isPlaybackMode()) {
           this.variables = rsp.bodyAsText ? JSON.parse(rsp.bodyAsText) : {};
         }
-        if (ensureExistence(this.sanitizer, "TestProxyHttpClient.sanitizer")) {
-          // Setting the recordingId in the sanitizer,
-          // the sanitizers added will take the recording id and only be part of the current test
-          this.sanitizer.setRecordingId(this.recordingId);
-          await handleEnvSetup(options.envSetupForPlayback, this.sanitizer);
-        }
+
+        await handleEnvSetup(
+          this.httpClient,
+          this.url,
+          this.recordingId,
+          options.envSetupForPlayback
+        );
+
         // Sanitizers to be added only in record mode
         if (isRecordMode() && options.sanitizerOptions) {
           // Makes a call to the proxy-tool to add the sanitizers for the current recording id
@@ -177,7 +180,7 @@ export class Recorder {
     this.stateManager.state = "stopped";
     if (this.recordingId !== undefined) {
       const stopUri = `${this.url}${isPlaybackMode() ? paths.playback : paths.record}${paths.stop}`;
-      const req = this._createRecordingRequest(stopUri);
+      const req = createRecordingRequest(stopUri, undefined, this.recordingId);
       req.headers.set("x-recording-save", "true");
 
       if (isRecordMode()) {
@@ -211,19 +214,16 @@ export class Recorder {
     }
   }
 
-  /**
-   * Adds the recording file and the recording id headers to the requests that are sent to the proxy tool.
-   * These are required to appropriately save the recordings in the record mode and picking them up in playback.
-   */
-  private _createRecordingRequest(url: string, method: HttpMethods = "POST") {
-    const req = createPipelineRequest({ url, method });
-    if (ensureExistence(this.sessionFile, "sessionFile")) {
-      req.body = JSON.stringify({ "x-recording-file": this.sessionFile });
+  async transformsInfo(): Promise<string | null | undefined> {
+    if (isLiveMode()) {
+      throw new RecorderError("Cannot call transformsInfo in live mode");
     }
-    if (this.recordingId !== undefined) {
-      req.headers.set("x-recording-id", this.recordingId);
+
+    if (ensureExistence(this.httpClient, "this.httpClient")) {
+      return await transformsInfo(this.httpClient, this.url, this.recordingId!);
     }
-    return req;
+
+    throw new RecorderError("Expected httpClient to be defined");
   }
 
   /**